requirements.txt
Django==3.2.6
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
psycopg2>=2.8
python-dotenv==0.19.0
django-cors-headers==3.5.0
requests

